root-198: implement autocomplete as text search option based on config
