+++ src/styles.css
@@ -1154,6 +1154,9 @@ textarea {
   @media (min-width: 900px) {
     .search-form__input {
       padding: .35rem; } }
+  .search-form__input:focus {
+    border-color: #333;
+    outline: none; }
 
 .search-form__submit {
   padding: .5rem .7rem;
@@ -1174,6 +1177,81 @@ textarea {
     height: 1em;
     width: 1em;
     fill: #fff; }
+
+/**
+ * autocomplete.scss
+ * Define autocomplete input + suggestion styles.
+ * @see: https://github.com/moroshko/react-autosuggest#theme-optional
+ *
+ * @copyright Copyright (c) 2017 Palantir.net
+ */
+.search-form__autocomplete-container {
+  display: flex; }
+  @media (min-width: 600px) {
+    .search-form__autocomplete-container {
+      width: 75%; } }
+  @media (min-width: 1100px) {
+    .search-form__autocomplete-container {
+      width: 50%; } }
+  .search-form__autocomplete-container .search-form__input-wrapper {
+    width: 100%; }
+
+.react-autosuggest__container {
+  position: relative;
+  flex: 1; }
+
+.react-autosuggest__input {
+  font-size: 0.8em;
+  line-height: 1.46667em;
+  flex: 1;
+  border: 1px solid #e5e5e5;
+  border-top-left-radius: 4px;
+  border-bottom-left-radius: 4px;
+  height: 34px;
+  padding: .5rem; }
+  .react-autosuggest__input:focus {
+    border-color: #333;
+    outline: none; }
+  .react-autosuggest__container--open .react-autosuggest__input {
+    border-bottom-left-radius: 0;
+    border-bottom-right-radius: 0; }
+
+.react-autosuggest__suggestions-container {
+  display: none; }
+
+.react-autosuggest__suggestions-container--open {
+  display: block;
+  position: absolute;
+  top: 33px;
+  width: 100%;
+  border: 1px solid #333;
+  background-color: #fff;
+  border-bottom-left-radius: 4px;
+  border-bottom-right-radius: 4px;
+  z-index: 2; }
+
+.react-autosuggest__suggestions-list {
+  margin: 0;
+  padding: 0;
+  list-style-type: none; }
+
+.react-autosuggest__suggestion {
+  font-size: 0.8em;
+  line-height: 1.46667em;
+  cursor: pointer;
+  padding: 10px 20px;
+  background-color: #fff;
+  border: 1px solid #fff; }
+  .react-autosuggest__suggestion:last-child {
+    border-bottom-left-radius: 4px;
+    border-bottom-right-radius: 4px; }
+
+.react-autosuggest__suggestion-link {
+  color: #737373; }
+
+.react-autosuggest__suggestion--highlighted {
+  background-color: #f6f6f6;
+  border: 1px solid #f6f6f6; }
 
 /**
  * search-form.scss
